requirements.txt
aiofiles==23.1.0
asgiref==3.7.2
asttokens==2.4.0
atomicwrites==1.4.0
attrs==21.4.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.13.0
backcall==0.2.0
certifi==2022.5.18.1
cffi==1.17.0
channels==4.1.0
charset-normalizer==2.0.12
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
cursor==1.3.5
daphne==4.1.2
debug==0.3.2
decorator==5.1.1
dict-toolbox==3.0.3
distlib==0.3.6
Django==4.2.6
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-jazzmin==2.6.0
django-money==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
enum34==1.1.10
exceptiongroup==1.1.3
executing==2.0.0
Faker==19.13.0
filelock==3.12.2
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.3
incremental==24.7.2
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.13
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msgpack==1.0.5
nest-asyncio==1.5.6
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.5.3
pluggy==1.0.0
pop==23.1.0
pop-config==12.0.1
pop-loop==1.0.6
prod==0.7
prompt-toolkit==3.0.39
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
py==1.11.0
py-moneyed==3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydevd-pycharm==212.5457.59
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.0.9
pyTelegramBotAPI==0.3.0
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rend==6.5.2
requests==2.27.1
rest-framework-simplejwt==0.0.2
see==1.4.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.11.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.23.0
wcwidth==0.2.8
zope.interface==7.0.3
